+++ ts/src/lib/actions/define-function2.ts
@@ -1,12 +1,12 @@
 import { $Boolean } from "kryo/builtins/boolean";
-import { $Bytes } from "kryo/builtins/bytes";
+import { $Uint16 } from "kryo/builtins/uint16";
 import { CaseStyle } from "kryo/case-style";
 import { ArrayType } from "kryo/types/array";
 import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { IntegerType } from "kryo/types/integer";
 import { LiteralType } from "kryo/types/literal";
 import { Ucs2StringType } from "kryo/types/ucs2-string";
-import { UintSize } from "semantic-types";
+import { Uint16, UintSize } from "semantic-types";
 import { ActionBase } from "../action-base";
 import { $ActionType, ActionType } from "../action-type";
 import { $Parameter, Parameter } from "../parameter";
@@ -25,7 +25,7 @@ export interface DefineFunction2 extends ActionBase {
   preloadGlobal: boolean;
   registerCount: UintSize;
   parameters: Parameter[];
-  body: Uint8Array;
+  bodySize: Uint16;
 }
 
 export const $DefineFunction2: DocumentIoType<DefineFunction2> = new DocumentType<DefineFunction2>(() => ({
@@ -48,7 +48,7 @@ export const $DefineFunction2: DocumentIoType<DefineFunction2> = new DocumentTyp
     preloadGlobal: {type: $Boolean},
     registerCount: {type: new IntegerType()},
     parameters: {type: new ArrayType({itemType: $Parameter, maxLength: Infinity})},
-    body: {type: $Bytes},
+    bodySize: {type: $Uint16},
   },
   changeCase: CaseStyle.SnakeCase,
 }));