+++ ts/src/lib/actions/try.ts
@@ -1,26 +1,27 @@
-import { $Bytes } from "kryo/builtins/bytes";
+import { $Uint16 } from "kryo/builtins/uint16";
 import { CaseStyle } from "kryo/case-style";
 import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { LiteralType } from "kryo/types/literal";
+import { Uint16 } from "semantic-types";
 import { ActionBase } from "../action-base";
 import { $ActionType, ActionType } from "../action-type";
 import { $CatchTarget, CatchTarget } from "../catch-target";
 
 export interface Try extends ActionBase {
   action: ActionType.Try;
-  try: Uint8Array;
-  catch?: Uint8Array;
+  trySize: Uint16;
+  catchSize?: Uint16;
   catchTarget: CatchTarget;
-  finally?: Uint8Array;
+  finallySize?: Uint16;
 }
 
 export const $Try: DocumentIoType<Try> = new DocumentType<Try>(() => ({
   properties: {
     action: {type: new LiteralType({type: $ActionType, value: ActionType.Try as ActionType.Try})},
-    try: {type: $Bytes},
-    catch: {type: $Bytes, optional: true},
+    trySize: {type: $Uint16},
+    catchSize: {type: $Uint16, optional: true},
     catchTarget: {type: $CatchTarget},
-    finally: {type: $Bytes, optional: true},
+    finallySize: {type: $Uint16, optional: true},
   },
   changeCase: CaseStyle.SnakeCase,
 }));